refcator: use my-batis-plus

+++ back-end/src/main/java/io/thescenery/shopAssist/user/controller/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
   }
 
   @GetMapping("/user/{id}")
-  UserInfoVo getUserById(@PathVariable int id) {
+  UserInfoVo getUserById(@PathVariable Long id) {
     return userService.getUserById(id);
   }
